rename user interface add entity access
and change user interface for user log in from social media
